requirements.txt
langfuse
streamlit
instructor
google-genai
openai
python-dotenv